#1216 show diagnostic signs on all folders
